requirements.txt
django
Pillow
django-ckeditor
